refactor(transactions): extract shared formatNum/parseNum helpers
