+++ frontend/src/components/transactions/numberFormat.ts
@@ -1,0 +1,12 @@
+export function formatNum(value: number | string): string {
+  if (value === '' || value === null || value === undefined) return ''
+  const num = typeof value === 'string' ? parseFloat(value) : value
+  if (isNaN(num)) return ''
+  const fixed = num.toFixed(2)
+  const [int, dec] = fixed.split('.')
+  return `${int.replace(/\B(?=(\d{3})+(?!\d))/g, ',')}.${dec}`
+}
+
+export function parseNum(value: string): number {
+  return parseFloat(value.replace(/,/g, '')) || 0
+}
