+++ frontend/src/pages/purchasing/CreatePurchaseOrderPage.tsx
@@ -43,6 +43,7 @@ import {
 } from '@/store/api/purchasingApi'
 import { useGetDocumentNumberSettingsQuery } from '@/store/api/settingsApi'
 import { LINE_ITEM_TABLE_SX } from '@/components/transactions/transactionTableStyles'
+import { formatNum, parseNum } from '@/components/transactions/numberFormat'
 import { formatCurrency, getCurrentDate, toMuiDatePickerFormat } from '@/utils/formatters'
 import { rtkErrorMessage } from '@/utils/errorMessage'
 
@@ -112,19 +113,6 @@ const emptyItem = (): PurchaseOrderItem => ({
   discountPercent: 0,
   totalPrice: 0,
 })
-
-function formatNum(value: number | string): string {
-  if (value === '' || value === null || value === undefined) return ''
-  const num = typeof value === 'string' ? parseFloat(value) : value
-  if (isNaN(num)) return ''
-  const fixed = num.toFixed(2)
-  const [int, dec] = fixed.split('.')
-  return `${int.replace(/\B(?=(\d{3})+(?!\d))/g, ',')}.${dec}`
-}
-
-function parseNum(value: string): number {
-  return parseFloat(value.replace(/,/g, '')) || 0
-}
 
 const COL_COUNT = 4
